organize interface files
