+++ src/bookshelf/index.ts
@@ -1,10 +1,9 @@
 import { assign, identity } from 'lodash';
 import { pluralize as plural } from 'inflection';
-import { SerialOpts, Serializer } from '../jsonapi-serializer';
-import { Mapper, MapOpts } from '../interfaces';
-import { Data, BookOpts } from './extras';
-import { LinkOpts } from '../links';
 
+import { SerialOpts, Serializer } from '../serializer';
+import { Mapper, MapOpts, LinkOpts } from '../interfaces';
+import { Data, BookOpts } from './extras';
 import { Information, processData, toJSON } from './utils';
 
 /**
