+++ src/bookshelf/links.ts
@@ -3,8 +3,8 @@ import { pluralize as plural } from 'inflection';
 import { stringify as queryParams } from 'qs';
 
 import { Model } from './extras';
-import { LinkOpts, PagOpts, QueryOpts } from '../links';
-import { LinkObj } from '../jsonapi-serializer';
+import { LinkOpts, PagOpts, QueryOpts } from '../interfaces';
+import { LinkObj } from '../serializer';
 
 function urlConcat(...parts: string[]): string {
   return parts.join('/');
