+++ src/interfaces/common.ts
