+++ src/interfaces/index.ts
@@ -1,0 +1,3 @@
+export * from './common';
+export * from './links';
+export * from './relations';
