+++ src/interfaces/links.ts
