+++ src/serializer/jsonapi-serializer.skel.d.ts
