+++ tsconfig.json
@@ -11,7 +11,6 @@
     "noImplicitReturns": true,
     "noImplicitThis": true
   },
-  "exclude": [
-    "node_modules"
-  ]
+  "include": ["src"],
+  "exclude": ["node_modules"]
 }
